requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
accelerate=0.21.0=pypi_0
addict=2.4.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bitsandbytes=0.43.1=pypi_0
bleach=6.1.0=pypi_0
bs4=0.0.2=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2022.12.7=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
datasets=2.18.0=pypi_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
deepspeed=0.12.6=pypi_0
defusedxml=0.7.1=pypi_0
diffusers=0.28.0.dev0=pypi_0
dill=0.3.8=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.6.1=pypi_0
einops-exts=0.0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.110.0=pypi_0
fastjsonschema=2.19.1=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.9.0=pypi_0
fonttools=4.50.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
ftfy=6.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gradio=4.16.0=pypi_0
gradio-client=0.8.1=pypi_0
grpcio=1.62.1=pypi_0
h11=0.14.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.0=pypi_0
huggingface-hub=0.22.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib-resources=6.4.0=pypi_0
importlib_metadata=7.1.0=hd8ed1ab_0
ipykernel=6.29.3=pyhd33586a_0
ipython=8.18.1=pyh707e725_3
ipywidgets=8.1.2=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.4.0=pypi_0
json5=0.9.24=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.4=pypi_0
jupyter-server=2.13.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py39hf3d152e_0
jupyterlab=4.1.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.4=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libnsl=2.0.1=hd590300_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.45.2=h2797004_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
lightning-utilities=0.11.2=pypi_0
llava=1.2.2.post1=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markdown2=2.4.13=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mmcv=1.7.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.3=pypi_0
nbformat=5.10.3=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
notebook=7.1.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.24.1=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
open-clip-torch=2.24.0=pypi_0
opencv-python=4.9.0.80=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.2.1=hd590300_1
optimum=1.18.0=pypi_0
orjson=3.9.15=pypi_0
overrides=7.7.0=pypi_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.1=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
peft=0.6.2=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=24.0=pypi_0
platformdirs=4.2.0=pyhd8ed1ab_0
pretty-errors=1.2.25=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=3.20.2=pypi_0
psutil=5.9.8=py39hd1e30aa_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=15.0.2=pypi_0
pyarrow-hotfix=0.6=pypi_0
pycparser=2.22=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pynvml=11.5.0=pypi_0
pyparsing=3.1.2=pypi_0
python=3.9.18=h0755675_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.7=pypi_0
python-multipart=0.0.9=pypi_0
python_abi=3.9=4_cp39
pytorch-fid=0.3.0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py39h6a678d5_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.34.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
ruff=0.4.2=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.13.0=pypi_0
semantic-version=2.10.0=pypi_0
send2trash=1.8.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=2.0.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py39h06a4308_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.38.2=hc218d9a_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.36.3=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.12=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.4.0=pypi_0
timm=0.6.13=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.15.1=pypi_0
tomli=2.0.1=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=2.1.2=pypi_0
torchaudio=2.1.2+cu118=pypi_0
torchmetrics=1.4.0=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.4=py39hd1e30aa_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pyhd8ed1ab_0
transformers=4.37.2=pypi_0
triton=2.1.0=pypi_0
typer=0.11.0=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.8.0=pypi_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=1.26.13=pypi_0
uvicorn=0.29.0=pypi_0
wandb=0.16.6=pypi_0
wavedrom=2.0.3.post3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39h06a4308_0
widgetsnbextension=4.0.10=pypi_0
xformers=0.0.23=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_0
yapf=0.40.1=pypi_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.18.1=pypi_0
zlib=1.2.13=hd590300_5
